requirements.txt
account==0.1.0
alembic==1.13.1
amqp==2.6.1
anyio==3.7.1
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==4.1.2
billiard==3.6.4.0
black==23.12.1
cachetools==5.3.2
celery==4.4.7
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.9.0
decorators==2.0.7
Django==5.0.1
django-appconf==1.0.6
django-debug-toolbar==4.2.0
django-ipware==6.0.3
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
djangorestframework==3.14.0
djangorestframework-recaptcha==0.2.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-spectacular==0.27.0
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
Faker==22.1.0
fastapi==0.70.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.9.0
idna==3.6
inflect==5.6.2
inflection==0.5.1
Jinja2==2.11.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==4.6.11
lxml==5.1.0
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.2.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pytest==5.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ipware==2.0.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3.post1
PyYAML==6.0.1
raven==6.10.0
recaptcha==1.0rc1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.11.8
vine==1.3.0
wcwidth==0.2.13
websockets==8.1
